dags/requirements.txt
apache-airflow==2.10.2
pendulum
pandas
minio
datetime
fastavro